Making things more cohesive
