+++ worsecrossbars/utilities/Logging.py
@@ -5,19 +5,16 @@ from worsecrossbars import configs
 
 class Logging:
 
-    def __init__ (self, number_hidden_layers, fault_type, number_ANNs, number_simulations):
+    def __init__ (self, args):
         """
 
         """
 
         log_number = 1
-        for _ in glob.glob(str(configs.working_dir.joinpath("outputs", "logs", f"spruce_faultType{fault_type}_{number_hidden_layers}HL-?.log"))):
+        for _ in glob.glob(str(configs.working_dir.joinpath("outputs", "logs", f"spruce_faultType{args.fault_type}_{args.number_hidden_layers}HL-?.log"))):
             log_number += 1
-        self.file_object = str(configs.working_dir.joinpath("outputs", "logs", f"spruce_faultType{fault_type}_{number_hidden_layers}HL-{log_number}.log"))
-        self.hidden_layers = number_hidden_layers
-        self.fault_type = fault_type
-        self.number_ANNs = number_ANNs
-        self.number_simulations = number_simulations
+        self.file_object = str(configs.working_dir.joinpath("outputs", "logs", f"spruce_faultType{args.fault_type}_{args.number_hidden_layers}HL-{log_number}.log"))
+        self.args = args
 
 
 
@@ -28,7 +25,7 @@ class Logging:
 
         if special == "begin":
             with open(self.file_object, 'a') as file:
-                file.write(f"----- Begin log {datetime.now().__str__()} -----\nAttempting simulation with following parameters:\nnumber_hidden_layers: {self.hidden_layers}\nfault_type: {self.fault_type}\nnumber_ANNs: {self.number_ANNs}\nnumber_simulations: {self.number_simulations}\n\n")
+                file.write(f"----- Begin log {datetime.now().__str__()} -----\nAttempting simulation with following parameters:\nnumber_hidden_layers: {self.args.hidden_layers}\nfault_type: {self.args.fault_type}\nnumber_ANNs: {self.args.number_ANNs}\nnumber_simulations: {self.args.number_simulations}\n\n")
         elif special == "end":
             with open(self.file_object, 'a') as file:
                 file.write(f"[{datetime.now().strftime('%H:%M:%S')}] Saved accuracies to file. Ending.\n----- End log {datetime.now().__str__()} -----")
